requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
discord-ext-voice_recv==0.5.2a179
discord.py==2.6.4
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
py-cord==2.6.1
pycparser==2.23
PyNaCl==1.5.0
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
webrtcvad==2.0.10
yarl==1.22.0
